+++ src/components/sidebar/tabs/queue/MediaLightbox.vue
@@ -85,14 +85,14 @@ import { computed, nextTick, ref, watch } from 'vue'
 
 import ComfyImage from '@/components/common/ComfyImage.vue'
 import Button from '@/components/ui/button/Button.vue'
-import type { AugmentedResultItem } from '@/stores/resultItem'
+import type { AugmentedResultItem } from '@/utils/resultItem'
+import { resultItemUrl } from '@/utils/resultItemUrl'
 import {
   isAudioResult,
   isImageResult,
   isTextResult,
-  isVideoResult,
-  resultItemUrl
-} from '@/stores/resultItem'
+  isVideoResult
+} from '@/utils/resultItem'
 
 import ResultAudio from './ResultAudio.vue'
 import ResultText from './ResultText.vue'
