+++ src/components/sidebar/tabs/queue/ResultText.vue
@@ -11,8 +11,8 @@
 
 <script setup lang="ts">
 import { useTextFileContent } from '@/composables/useTextFileContent'
-import type { AugmentedResultItem } from '@/stores/resultItem'
-import { resultItemUrl } from '@/stores/resultItem'
+import type { AugmentedResultItem } from '@/utils/resultItem'
+import { resultItemUrl } from '@/utils/resultItemUrl'
 
 const { result } = defineProps<{
   result: AugmentedResultItem
