+++ src/composables/queue/useJobMenu.ts
@@ -18,13 +18,9 @@ import { useExecutionStore } from '@/stores/executionStore'
 import { useNodeDefStore } from '@/stores/nodeDefStore'
 import { useQueueStore } from '@/stores/queueStore'
 import type { TaskItemImpl } from '@/stores/queueStore'
-import type { AugmentedResultItem } from '@/stores/resultItem'
-import {
-  isAudioResult,
-  isImageResult,
-  isVideoResult,
-  resultItemUrl
-} from '@/stores/resultItem'
+import type { AugmentedResultItem } from '@/utils/resultItem'
+import { resultItemUrl } from '@/utils/resultItemUrl'
+import { isAudioResult, isImageResult, isVideoResult } from '@/utils/resultItem'
 import { createAnnotatedPath } from '@/utils/createAnnotatedPath'
 import { appendJsonExt } from '@/utils/formatUtil'
 import { isResultItemType } from '@/utils/typeGuardUtil'
