+++ src/platform/assets/composables/useMediaAssetGalleryStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from 'pinia'
 import { ref, shallowRef } from 'vue'
 
-import type { AugmentedResultItem } from '@/stores/resultItem'
+import type { AugmentedResultItem } from '@/utils/resultItem'
 
 import type { AssetMeta } from '../schemas/mediaAssetSchema'
