+++ src/renderer/extensions/linearMode/flattenNodeOutput.ts
@@ -1,6 +1,6 @@
 import type { NodeExecutionOutput } from '@/schemas/apiSchema'
 import { parseNodeOutput } from '@/stores/resultItemParsing'
-import type { AugmentedResultItem } from '@/stores/resultItem'
+import type { AugmentedResultItem } from '@/utils/resultItem'
 
 export function flattenNodeOutput([nodeId, nodeOutput]: [
   string | number,
