+++ src/renderer/extensions/linearMode/useOutputHistory.ts
@@ -16,7 +16,7 @@ import { useAppModeStore } from '@/stores/appModeStore'
 import { useCommandStore } from '@/stores/commandStore'
 import { useExecutionStore } from '@/stores/executionStore'
 import { useQueueStore } from '@/stores/queueStore'
-import type { AugmentedResultItem } from '@/stores/resultItem'
+import type { AugmentedResultItem } from '@/utils/resultItem'
 
 export function useOutputHistory(): {
   outputs: PagedList<AssetItem>
