+++ src/stores/queueStore.ts
@@ -9,8 +9,8 @@ import type {
 } from '@/platform/remote/comfyui/jobs/jobTypes'
 import type { StatusWsMessageStatus, TaskOutput } from '@/schemas/apiSchema'
 import { api } from '@/scripts/api'
-import type { AugmentedResultItem } from '@/stores/resultItem'
-import { filterPreviewableResults } from '@/stores/resultItem'
+import type { AugmentedResultItem } from '@/utils/resultItem'
+import { filterPreviewableResults } from '@/utils/resultItem'
 import { parseTaskOutput } from '@/stores/resultItemParsing'
 import type { ComfyApp } from '@/scripts/app'
 import { useExtensionService } from '@/services/extensionService'
