+++ src/stores/resultItemParsing.test.ts
@@ -4,7 +4,7 @@ import { describe, expect, it, vi } from 'vitest'
 vi.mock('@/platform/assets/composables/media/assetMappers')
 
 import type { NodeExecutionOutput } from '@/schemas/apiSchema'
-import { isTextResult, resultItemSupportsPreview } from '@/stores/resultItem'
+import { isTextResult, resultItemSupportsPreview } from '@/utils/resultItem'
 import { parseNodeOutput, parseTaskOutput } from '@/stores/resultItemParsing'
 
 function makeOutput(
