+++ src/utils/resultItem.ts
@@ -1,6 +1,4 @@
-import { appendCloudResParam } from '@/platform/distribution/cloudPreviewUtil'
 import type { ResultItem } from '@/schemas/apiSchema'
-import { api } from '@/scripts/api'
 import type { SerializedNodeId } from '@/types/nodeId'
 import { getMediaTypeFromFilename } from '@/utils/formatUtil'
 
@@ -14,40 +12,6 @@ export interface AugmentedResultItem extends ResultItem {
   frame_rate?: number
   url?: string
   previewUrl?: string
-}
-
-function resultItemUrlParams(item: AugmentedResultItem): URLSearchParams {
-  const params = new URLSearchParams()
-  params.set('filename', item.filename ?? '')
-  params.set('type', item.type ?? '')
-  params.set('subfolder', item.subfolder ?? '')
-  if (item.format) params.set('format', item.format)
-  if (item.frame_rate) params.set('frame_rate', item.frame_rate.toString())
-  return params
-}
-
-export function resultItemUrl(item: AugmentedResultItem): string {
-  if (item.url !== undefined) return item.url
-  if (!item.filename) return ''
-  return api.apiURL('/view?' + resultItemUrlParams(item))
-}
-
-export function resultItemPreviewUrl(item: AugmentedResultItem): string {
-  if (item.previewUrl !== undefined) return item.previewUrl
-  if (!isImageResult(item)) return resultItemUrl(item)
-  const params = resultItemUrlParams(item)
-  appendCloudResParam(params, item.filename)
-  return api.apiURL('/view?' + params)
-}
-
-export function resultItemVhsAdvancedPreviewUrl(
-  item: AugmentedResultItem
-): string {
-  return api.apiURL('/viewvideo?' + resultItemUrlParams(item))
-}
-
-export function resultItemUrlWithTimestamp(item: AugmentedResultItem): string {
-  return `${resultItemUrl(item)}&t=${+new Date()}`
 }
 
 const isVideoBySuffix = (item: AugmentedResultItem): boolean =>
@@ -136,13 +100,4 @@ export function filterPreviewableResults(
   items: readonly AugmentedResultItem[]
 ): AugmentedResultItem[] {
   return items.filter(resultItemSupportsPreview)
-}
-
-export function findResultIndexByUrl(
-  items: readonly AugmentedResultItem[],
-  url?: string
-): number {
-  if (!url) return 0
-  const idx = items.findIndex((item) => resultItemUrl(item) === url)
-  return idx >= 0 ? idx : 0
 }
